+++ pkg/observable/channel/map.go
@@ -2,7 +2,6 @@ package channel
 
 import (
 	"context"
-	"sync"
 
 	"github.com/pokt-network/poktroll/pkg/observable"
 )
@@ -104,31 +103,6 @@ func ForEach[V any](
 			return zeroValue[V](), true
 		},
 	)
-}
-
-// Collect collects all notifications received from the observable and returns
-// them as a slice. Either srcObservable MUST be bounded (i.e. known to close),
-// OR ctx MUST be canceled, after some finite duration as it blocks until either
-// srcObservable is closed OR ctx is canceled. Collect is a terminal observable
-// operator.
-func Collect[V any](
-	ctx context.Context,
-	srcObservable observable.Observable[V],
-) (dstCollection []V) {
-	var dstCollectionMu sync.Mutex
-	defer dstCollectionMu.Unlock()
-
-	ForEach(ctx, srcObservable, func(ctx context.Context, src V) {
-		dstCollectionMu.Lock()
-		dstCollection = append(dstCollection, src)
-		dstCollectionMu.Unlock()
-	})
-
-	// Wait for context to be done before returning.
-	<-ctx.Done()
-
-	dstCollectionMu.Lock()
-	return dstCollection
 }
 
 // goMapTransformNotification transforms, optionally skips, and publishes
