+++ pkg/relayer/miner/relay_fixtures_test.go
@@ -7,7 +7,7 @@ package miner_test
 
 var (
 	// marshaledMinableRelaysHex are the hex encoded strings of serialized
-	// relayer.MinedRelays which have been pre-mined to difficulty 2 by
+	// relayer.MinedRelays which have been pre-mined to difficulty 16 by
 	// populating the signature with random bytes. It is intended for use
 	// in tests.
 	marshaledMinableRelaysHex = []string{
